Migrating Classification Groups
In order for other extensions to use Neutron Classifier's
Classification Groups extensions will need to make a foreign
key association to the classification group's id.
When the DB migrations are being run the Classification Group
table would need to already exist before it can be referenced.
As Neutron runs it's own DB migrations before other extensions
this means that the QoS extensions would not be able to use
Classifications.
This patch is part of 2 patches, one to remove the classification
groups from Neutron Classifiers migrations and another to insert
it into Neutron's migrations.

Depends-on: I03133e2213aae07d1e5a47ee845595dad43b638e
Change-Id: Ibf5424b643b027da1fd03780f7ef81b970400c28